refactor front-end Looper into a component
